+++ src/main/java/com/medicine/medicine/dto/UserDTO.java
@@ -1,0 +1,18 @@
+package com.medicine.medicine.dto;
+public class UserDTO {
+    private Integer count;
+
+    public UserDTO() {}
+
+    public UserDTO(Integer count) {
+        this.count = count;
+    }
+
+    public Integer getCount() {
+        return count;
+    }
+
+    public void setCount(Integer count) {
+        this.count = count;
+    }
+}
